Extract product specific builders to separate objects
