+++ support-workers/src/main/scala/com/gu/zuora/subscriptionBuilders/ProductSubscriptionBuilders.scala
@@ -1,0 +1,95 @@
+package com.gu.zuora.subscriptionBuilders
+
+import java.util.UUID
+
+import com.gu.i18n.Country
+import com.gu.support.catalog
+import com.gu.support.catalog.{ProductRatePlan, ProductRatePlanId}
+import com.gu.support.config.ZuoraConfig
+import com.gu.support.promotions.{PromoCode, PromotionService}
+import com.gu.support.workers._
+import com.gu.support.workers.exceptions.CatalogDataNotFoundException
+import com.gu.support.zuora.api.ReaderType.Direct
+import com.gu.support.zuora.api._
+import org.joda.time.{DateTimeZone, Days, LocalDate}
+
+object ProductSubscriptionBuilders {
+
+  def validateRatePlan(maybeProductRatePlan: Option[ProductRatePlan[catalog.Product]], productDescription: String): ProductRatePlanId =
+    maybeProductRatePlan.map(_.id) match {
+      case Some(value) => value
+      case None => throw new CatalogDataNotFoundException(s"RatePlanId not found for $productDescription")
+    }
+
+  def buildContributionSubscription(contribution: Contribution, requestId: UUID, config: ZuoraConfig): SubscriptionData = {
+    val contributionConfig = config.contributionConfig(contribution.billingPeriod)
+    buildProductSubscription(
+      requestId,
+      contributionConfig.productRatePlanId,
+      List(
+        RatePlanChargeData(
+          ContributionRatePlanCharge(contributionConfig.productRatePlanChargeId, price = contribution.amount) //Pass the amount the user selected into Zuora
+        )
+      ),
+      readerType = Direct
+    )
+  }
+
+  def buildProductSubscription(
+    createdRequestId: UUID,
+    productRatePlanId: ProductRatePlanId,
+    ratePlanCharges: List[RatePlanChargeData] = Nil,
+    contractEffectiveDate: LocalDate = LocalDate.now(DateTimeZone.UTC),
+    contractAcceptanceDate: LocalDate = LocalDate.now(DateTimeZone.UTC),
+    readerType: ReaderType,
+    initialTermMonths: Int = 12
+  ): SubscriptionData = {
+    val (initialTerm, autoRenew, initialTermPeriodType) = if (readerType == ReaderType.Gift)
+      (initialTermInDays(contractEffectiveDate, contractAcceptanceDate, initialTermMonths), false, Day)
+    else
+      (12, true, Month)
+
+    SubscriptionData(
+      List(
+        RatePlanData(
+          RatePlan(productRatePlanId),
+          ratePlanCharges,
+          Nil
+        )
+      ),
+      Subscription(
+        contractEffectiveDate = contractEffectiveDate,
+        contractAcceptanceDate = contractAcceptanceDate,
+        termStartDate = contractEffectiveDate,
+        createdRequestId__c = createdRequestId.toString,
+        readerType = readerType,
+        autoRenew = autoRenew,
+        initialTerm = initialTerm,
+        initialTermPeriodType = initialTermPeriodType,
+      )
+    )
+  }
+
+  def initialTermInDays(contractEffectiveDate: LocalDate, contractAcceptanceDate: LocalDate, termLengthMonths: Int): Int = {
+    val termEnd = contractAcceptanceDate.plusMonths(termLengthMonths)
+    Days.daysBetween(contractEffectiveDate, termEnd).getDays
+  }
+
+  def applyPromoCode(
+    promotionService: PromotionService,
+    maybePromoCode: Option[PromoCode],
+    country: Country,
+    productRatePlanId: ProductRatePlanId,
+    subscriptionData: SubscriptionData
+  ) = {
+    val withPromotion = maybePromoCode.map { promoCode =>
+      for {
+        promotionWithCode <- promotionService.findPromotion(promoCode)
+        subscriptionWithPromotion <- promotionService.applyPromotion(promotionWithCode, country, productRatePlanId, subscriptionData, isRenewal = false)
+      } yield subscriptionWithPromotion
+    }
+
+    withPromotion.getOrElse(Right(subscriptionData))
+  }
+
+}
